+++ benchmarks/src/test/scala/org/ergoplatform/ModifiersApplicationBench.scala
@@ -2,7 +2,8 @@ package org.ergoplatform
 
 import org.ergoplatform.Utils.BenchReport
 import org.ergoplatform.modifiers.ErgoPersistentModifier
-import org.ergoplatform.modifiers.history.{BlockTransactions, Extension, Header}
+import org.ergoplatform.modifiers.history.extension.Extension
+import org.ergoplatform.modifiers.history.{BlockTransactions, Header}
 import org.ergoplatform.nodeView.{ErgoModifiersCache, NVBenchmark}
 import org.ergoplatform.nodeView.history.ErgoHistory
 import org.ergoplatform.nodeView.state.StateType
