+++ src/main/scala/org/ergoplatform/mining/CandidateBlock.scala
@@ -2,7 +2,8 @@ package org.ergoplatform.mining
 
 import io.circe.Encoder
 import io.circe.syntax._
-import org.ergoplatform.modifiers.history.{ExtensionCandidate, Header}
+import org.ergoplatform.modifiers.history.Header
+import org.ergoplatform.modifiers.history.extension.ExtensionCandidate
 import org.ergoplatform.modifiers.mempool.ErgoTransaction
 import org.ergoplatform.settings.Algos
 import scorex.core.block.Block
