+++ src/main/scala/org/ergoplatform/modifiers/ErgoFullBlock.scala
@@ -3,7 +3,8 @@ package org.ergoplatform.modifiers
 import io.circe.syntax._
 import io.circe.{Decoder, Encoder, HCursor, Json}
 import org.ergoplatform.http.api.ApiCodecs
-import org.ergoplatform.modifiers.history.{ADProofs, BlockTransactions, Extension, Header}
+import org.ergoplatform.modifiers.history.extension.Extension
+import org.ergoplatform.modifiers.history.{ADProofs, BlockTransactions, Header}
 import org.ergoplatform.modifiers.mempool.ErgoTransaction
 import scorex.core.serialization.ScorexSerializer
 import scorex.core.{ModifierTypeId, TransactionsCarryingPersistentNodeViewModifier}
