+++ src/main/scala/org/ergoplatform/modifiers/ErgoPersistentModifier.scala
@@ -1,7 +1,8 @@
 package org.ergoplatform.modifiers
 
 import io.circe.Encoder
-import org.ergoplatform.modifiers.history.{ADProofs, BlockTransactions, Extension, Header}
+import org.ergoplatform.modifiers.history.extension.Extension
+import org.ergoplatform.modifiers.history.{ADProofs, BlockTransactions, Header}
 import scorex.core.PersistentNodeViewModifier
 
 trait ErgoPersistentModifier extends PersistentNodeViewModifier with ErgoNodeViewModifier
