+++ src/main/scala/org/ergoplatform/modifiers/history/extension/Extension.scala
@@ -1,18 +1,17 @@
-package org.ergoplatform.modifiers.history
+package org.ergoplatform.modifiers.history.extension
 
 import com.google.common.primitives.Bytes
 import io.circe.syntax._
 import io.circe.{Decoder, Encoder, HCursor}
 import org.ergoplatform.http.api.ApiCodecs
 import org.ergoplatform.modifiers.BlockSection
-import org.ergoplatform.settings.{Algos, Constants}
+import org.ergoplatform.settings.Algos
 import scorex.core.ModifierTypeId
 import scorex.core.serialization.ScorexSerializer
 import scorex.crypto.authds.LeafData
-import scorex.crypto.authds.merkle.{Leaf, MerkleProof, MerkleTree}
+import scorex.crypto.authds.merkle.MerkleTree
 import scorex.crypto.hash.Digest32
-import scorex.util.serialization.{Reader, Writer}
-import scorex.util.{ModifierId, bytesToId, idToBytes}
+import scorex.util.ModifierId
 
 /**
   * Extension section of Ergo block. Contains key-value storage
@@ -37,28 +36,6 @@ case class Extension(headerId: ModifierId,
       s"fields: ${fields.map(kv => s"${Algos.encode(kv._1)} -> ${Algos.encode(kv._2)}")}) "
   }
 
-}
-
-/**
-  * Extension block section with not filled header id
-  */
-class ExtensionCandidate(val fields: Seq[(Array[Byte], Array[Byte])]) {
-  lazy val merkleTree: MerkleTree[Digest32] = Extension.merkleTree(fields)
-
-  lazy val digest: Digest32 = Algos.merkleTreeRoot(merkleTree)
-
-  def toExtension(headerId: ModifierId): Extension = Extension(headerId, fields)
-
-  def ++(that: ExtensionCandidate): ExtensionCandidate = ExtensionCandidate(fields ++ that.fields)
-
-  def proofFor(key: Array[Byte]): Option[MerkleProof[Digest32]] =
-    fields.find(_._1 sameElements key)
-      .map(Extension.kvToLeaf)
-      .flatMap(kv => merkleTree.proofByElement(Leaf[Digest32](LeafData @@ kv)(Algos.hash)))
-}
-
-object ExtensionCandidate {
-  def apply(fields: Seq[(Array[Byte], Array[Byte])]): ExtensionCandidate = new ExtensionCandidate(fields)
 }
 
 object Extension extends ApiCodecs {
@@ -94,33 +71,6 @@ object Extension extends ApiCodecs {
       fields <- c.downField("fields").as[List[(Array[Byte], Array[Byte])]]
     } yield Extension(headerId, fields)
   }
-}
-
-object ExtensionSerializer extends ScorexSerializer[Extension] {
-
-  override def serialize(obj: Extension, w: Writer): Unit = {
-    w.putBytes(idToBytes(obj.headerId))
-    w.putUShort(obj.fields.size)
-    obj.fields.foreach { case (key, value) =>
-      w.putBytes(key)
-      w.putUByte(value.length)
-      w.putBytes(value)
-    }
-  }
-
-  override def parse(r: Reader): Extension = {
-    val startPosition = r.position
-    val headerId = bytesToId(r.getBytes(Constants.ModifierIdSize))
-    val fieldsSize = r.getUShort()
-    val fieldsView = (1 to fieldsSize).toStream.map { _ =>
-      val key = r.getBytes(Extension.FieldKeySize)
-      val length = r.getUByte()
-      val value = r.getBytes(length)
-      (key, value)
-    }
-    val fields = fieldsView.takeWhile(_ => r.position - startPosition < Constants.MaxExtensionSizeMax)
-    require(r.position - startPosition < Constants.MaxExtensionSizeMax)
-    Extension(headerId, fields, Some(r.position - startPosition))
-  }
 
 }
+
