+++ src/test/scala/org/ergoplatform/nodeView/history/BlockSectionValidationSpecification.scala
@@ -2,6 +2,7 @@ package org.ergoplatform.nodeView.history
 
 import org.ergoplatform.modifiers.BlockSection
 import org.ergoplatform.modifiers.history._
+import org.ergoplatform.modifiers.history.extension.Extension
 import org.ergoplatform.nodeView.state.StateType
 import org.ergoplatform.utils.HistoryTestHelpers
 import scorex.core.block.Block.Version
