+++ src/test/scala/org/ergoplatform/nodeView/state/ErgoStateContextSpec.scala
@@ -1,7 +1,7 @@
 package org.ergoplatform.nodeView.state
 
 import org.ergoplatform.modifiers.ErgoFullBlock
-import org.ergoplatform.modifiers.history.Extension
+import org.ergoplatform.modifiers.history.extension.Extension
 import org.ergoplatform.modifiers.history.popow.NipopowAlgos
 import org.ergoplatform.settings.Parameters._
 import org.ergoplatform.utils.HistoryTestHelpers
